ActorSystem::Remove size_t arg from int

+++ VEngine/Code/Actors/ActorSystem.h
@@ -50,7 +50,7 @@ public:
 		return actor.get();
 	}
 
-	void Remove(int index)
+	void Remove(size_t index)
 	{
 		auto components = actors[index]->GetAllComponents();
 		for (int i = 0; i < components.size(); i++)
